--- week3_memo/src/main/java/com/likelion/memoapp/memo/dto/MemoUpdateRequestDTO.java
@@ -1,10 +1,0 @@
-package com.likelion.memoapp.memo.dto;
-
-import lombok.Getter;
-
-@Getter
-public class MemoUpdateRequestDTO extends RequestDTO {
-    private String title;
-    private String content;
-
-}
