--- week3_memo/src/main/java/com/likelion/memoapp/memo/dto/RequestDTO.java
@@ -1,10 +1,0 @@
-package com.likelion.memoapp.memo.dto;
-
-import java.util.UUID;
-import lombok.Getter;
-
-@Getter
-public class RequestDTO {
-    UUID userId;
-
-}
