+++ week3_memo/src/main/java/com/likelion/memoapp/model/dto/memo/MemoCreateRequestDTO.java
@@ -1,10 +1,10 @@
-package com.likelion.memoapp.memo.dto;
+package com.likelion.memoapp.model.dto.memo;
 
 import java.util.UUID;
 import lombok.Getter;
 
 @Getter
-public class MemoCreateRequestDTO extends RequestDTO {
+public class MemoCreateRequestDTO extends MemoRequestDTO {
     private UUID userId;
     private String title;
     private String content;
