+++ week3_memo/src/main/java/com/likelion/memoapp/model/dto/memo/MemoRequestDTO.java
@@ -1,0 +1,9 @@
+package com.likelion.memoapp.model.dto.memo;
+
+import java.util.UUID;
+import lombok.Getter;
+
+@Getter
+public class MemoRequestDTO {
+    UUID userId;
+}
