+++ week3_memo/src/main/java/com/likelion/memoapp/model/dto/memo/MemoUpdateRequestDTO.java
@@ -1,0 +1,10 @@
+package com.likelion.memoapp.model.dto.memo;
+
+import lombok.Getter;
+
+@Getter
+public class MemoUpdateRequestDTO extends MemoRequestDTO {
+    private String title;
+    private String content;
+
+}
