+++ week3_memo/src/main/java/com/likelion/memoapp/model/dto/user/UserRequestDTO.java
@@ -1,0 +1,8 @@
+package com.likelion.memoapp.model.dto.user;
+
+import lombok.Getter;
+
+@Getter
+public class UserRequestDTO {
+    private String name;
+}
