--- week3_memo/src/main/java/com/likelion/memoapp/user/dto/RequestDTO.java
@@ -1,8 +1,0 @@
-package com.likelion.memoapp.user.dto;
-
-import lombok.Getter;
-
-@Getter
-public class RequestDTO {
-    private String name;
-}
